--- a/.gitignore
+++ b/.gitignore
@@ -110,3 +110,6 @@ venv.bak/
 .mypy_cache/
 .dmypy.json
 dmypy.json
+
+# hatch-vcs
+src/pypistats/_version.py