--- a/pyproject.toml
+++ b/pyproject.toml
@@ -69,6 +69,9 @@ scripts.pypistats = "pypistats.cli:main"
 [tool.hatch]
 version.source = "vcs"
 
+[tool.hatch.build.hooks.vcs]
+version-file = "src/pypistats/_version.py"
+
 [tool.hatch.version.raw-options]
 local_scheme = "no-local-version"