With Python 3.12.4:

```sh
python -X importtime -c "import pypistats" 2> import.log && tuna import.log
```

Cuts out the 23ms `importlib.metadata` import.

# `main`

<img width="1624" alt="image" src="https://github.com/user-attachments/assets/4c8a1e9b-889b-4a0b-95cc-8c8366d49caf">

```console
❯ hyperfine --warmup 8 "pypistats --version"
Benchmark 1: pypistats --version
  Time (mean ± σ):      57.2 ms ±   3.3 ms    [User: 44.6 ms, System: 10.9 ms]
  Range (min … max):    51.7 ms …  72.8 ms    48 runs
```

# PR

<img width="1624" alt="image" src="https://github.com/user-attachments/assets/1a4166e2-59bc-4a00-a341-8347be91f9ac">


```console
❯ hyperfine --warmup 8 "pypistats --version"
Benchmark 1: pypistats --version
  Time (mean ± σ):      37.4 ms ±   2.7 ms    [User: 29.0 ms, System: 7.3 ms]
  Range (min … max):    34.6 ms …  55.4 ms    74 runs
```

---

And update some config.